Split oom-proc subtests out of test-io-opt
The sendfile/copy_file_range "rereads synthetic oom proc source"
subtests exercise copy_fd_range's own proc-source read-intercept
wiring, which has no Linux-kernel counterpart -- no real program
sendfile()s or copy_file_range()s out of procfs. Their presence in
test-io-opt forced the whole binary into test-matrix.sh's QEMU_SKIP
list, silently withholding the other 5 subtests (sendfile, fsync,
fallocate x2, copy_file_range) from ever being checked against the
real Alpine linux-virt kernel.

Move the two subtests to a new tests/test-oom-proc.c (make check
only, never invoked from test-matrix.sh) and drop test-io-opt from
QEMU_SKIP. Also fix an open-failure fallback both subtests carried
over from test-io-opt: a failed open() of the oom_adj/oom_score_adj
proc nodes fell through to PASS() instead of FAIL(), so if elfuse
ever stopped serving these synthetic files the test would report
green instead of catching the regression it exists to guard against.

Verified: make check green (test-io-opt + test-oom-proc both OK),
test-matrix.sh elfuse-aarch64 175/0, and test-matrix.sh qemu-aarch64
170/0 with test-io-opt now passing against the real kernel for the
first time.
